fix: add isUserCreated as action
