+++ routes/all_access/post.routes.js
@@ -3,6 +3,7 @@ const orderController = require("../../controllers/order.controller");
 const {validateIfBodyUndefined, orderValidators, verifySignUp, checkUserData} = require("../../validators");
 const masterController = require("../../controllers/master.controller");
 const controller = require("../../controllers/auth.controller");
+const authController = require("../../controllers/auth.controller");
 
 router.use(validateIfBodyUndefined);
 
@@ -22,5 +23,7 @@ router.post("/signin", controller.signin);
 
 router.post("/orders", orderValidators, orderController.create);
 router.post("/masters/available", masterController.findAllMastersAvailable);
+
+router.post('/verify/user/created', authController.isUserCreated);
 
 module.exports = router;
