+++ validators/orderValidators.js
@@ -25,52 +25,6 @@ ifDateTimeAppropriate = async (req, res, next) => {
     next();
 }
 
-ifUserCreated = async (req, res, next) => {
-    logger.info('Verifying if such user already exist...');
-    const userToFind = {
-        name: req.body.name,
-        email: req.body.email
-    }
-    try {
-        const [user, isUserCreated] = await User.findOrCreate({
-            where: userToFind,
-            defaults: userToFind
-        });
-        if (isUserCreated) {
-            await user.setRoles([1]);
-            logger.info('User has been created as new. Heading next...');
-        } else {
-            logger.info('Such user has been found. Checking authorization...');
-            const token = req.headers["x-access-token"];
-            const error = {};
-            jwt.verify(token, config.secret, async (err, decoded) => {
-                if (err) {
-                    logger.info('error');
-                    if (user.emailChecked) {
-                        error.message = 'Log in before placing new order!';
-                    } else {
-                        error.message = 'Your email is unchecked. Please check your email for confirmation letter';
-                    }
-                } else if (decoded?.id !== user.id) {
-                    error.message = "User authorized but name and email belong to another user";
-                }
-            });
-            if (error.message) {
-                logger.error(error.message);
-                return res.status(401).send({
-                    message: error.message,
-                    code: 401
-                });
-            }
-        }
-        req.body.userId = user.id;
-        next();
-    } catch (e) {
-        logger.error(e.message + ': Check user credentials.');
-        return res.status(400).send({message: e.message + ': Check user credentials.'});
-    }
-}
-
 ifOrderCanBePlaced = async (req, res, next) => {
     logger.info('Verifying order credentials...');
     const timeInNum = parseInt(req.body.time.substring(0, 2));
@@ -102,4 +56,4 @@ ifOrderCanBePlaced = async (req, res, next) => {
     }
 }
 
-module.exports = [ifDateTimeAppropriate, ifUserCreated, ifOrderCanBePlaced];
+module.exports = [ifDateTimeAppropriate, ifOrderCanBePlaced];
